LayoutTests/http/tests/websocket/tests/hixie76/bad-sub-protocol-empty-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket ''
Test WebSocket bad sub-protocol names (empty).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS successfullyParsed is true

TEST COMPLETE
